+++ typescript/CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Next (v0.2.0)
 
 ### `FrozenWeightedList`
+- New: `is_zero()` method
 - `.items()` correctly freezes returned objects
 - Fix off-by-1 errors in indexing for sampling without replacement
 - Improve error messages